splash WIP rename button to submit button
