--- frontend/src/components/Button.js
@@ -1,9 +1,0 @@
-export default function Button(props) {
-  return (
-    <button
-      className={`${props.className} font-bold my-4 px-3 py-1 rounded-full bg-yellow-300 shadow-sm`}
-    >
-      {props.title}
-    </button>
-  );
-}
