+++ frontend/src/components/SubmitButton.js
@@ -1,0 +1,9 @@
+export default function SubmitButton(props) {
+  return (
+    <button
+      className={`${props.className} text-xl font-bold my-4 px-4 py-2 rounded-full bg-yellow-300 shadow-sm`}
+    >
+      {props.title}
+    </button>
+  );
+}
