+++ frontend/src/index.css
@@ -9,3 +9,7 @@
 .menu-section {
   @apply py-1 bg-yellow-300 rounded-full font-bold;
 }
+
+.splash-background {
+  background: url(./images/splash-bg-1.jpg);
+}
